app/pyproject.toml
[tool.poetry]
name = "flask-docker-example"
version = "0.1.0"
description = "Flask dockerize example"
license = "MIT"
authors = [
    "João Maia <joao@joaovrmaia.com>"
]
readme = "README.md"
repository = "https://github.com/jvrmaia/flask-docker-example"
homepage = "https://github.com/jvrmaia/flask-docker-example"


[tool.poetry.dependencies]
python = "3.8.2"
click = "==7.0"
flask = "==1.1.1"
itsdangerous = "==1.1.0"
jinja2 = "==2.11.1"
markupsafe = "==1.1.1"
python-decouple = "==3.3"
werkzeug = "==1.0.0"
gunicorn = "^20.0.4"



[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
